# src/docs/swagger.yaml
openapi: 3.0.1
info:
  title: Insider API
  version: "1.0.0"
  description: >
    Full API documentation for the Insider backend. All routes are mounted under /api.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/payments/stripe/webhook:
    post:
      tags: [Payments]
      summary: Stripe webhook
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/payments/webhook:
    post:
      tags: [Payments]
      summary: Partner webhook
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GenericResponse"
...
  schemas:
    GenericRequest:
      type: object
      description: Generic request payload.
      additionalProperties: true
    GenericResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties: true
      additionalProperties: true
...
security:
  - bearerAuth: []